requirements.txt
streamlit
requests
python-dotenv
google-generativeai
PyPDF2
feedparser
faster-whisper
pydub
scipy
numpy